[#12] Moved `RoundToCents` to the util package
